Add: [Deployer] Check if we need to deploy a new version before we do
Helm does a poor job on this, and tends to redeploy sometimes
(but not always). So lets take over from him and do it ourself.

Helm has no real means to request the current appVersion of a
release. You have to do with with 'helm ls', as the expected
'helm status' does not include this information.