Fix: before watching, fetch list, to get a recent resource_version to watch from
Watch() only returns entries, not the list. Entries in a list
can be very old, and no longer valid to watch from. Additionally,
the initial list watch() returns is unordered.

To correct way of doing this is first list the entries. Here the
list itself contains a resource_version that is recent. Using this
version to watch will result in a correct watch().

A lot has been written about this. Most clear resource for me was:
https://github.com/kubernetes-client/python/issues/819